refactor: split in-built methods into src/runtime/in-built
Move each global builtin and its helpers out of the monolithic
globals.ts file into grouped modules. createGlobalEnv now only
registers constants and imported natives on the environment.

Closes #77
