+++ src/runtime/in-built/io/index.ts
@@ -1,0 +1,4 @@
+export { tangaza_amakuru } from './tangaza_amakuru';
+export { sisitemu } from './sisitemu';
+export { injiza_amakuru } from './injiza_amakuru';
+export { hagarara } from './hagarara';
